+++ db/CoinGecko/StoredProcedure/Stablecoin_GetAll.sql
@@ -1,0 +1,15 @@
+CREATE OR ALTER PROCEDURE [CoinGecko].[Stablecoin_GetAll]
+AS
+BEGIN
+  SET NOCOUNT ON;
+
+  SELECT
+    [Id],
+    [Symbol],
+    [Name],
+    [CurrentPrice],
+    [DateCreated],
+    [DateUpdated],
+    [DataDate]
+  FROM [CoinGecko].[Stablecoin];
+END;